requirements.txt
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
Django==1.11.4
django-appconf==1.0.2
django-braces==1.12.0
django-generic-positions==0.2.2
django-libs==1.67.13
django-model-utils==3.1.1
django-phonenumber-field==2.0.0
django-social-share==1.1.2
easy-thumbnails==2.5
html5lib==1.0b10
httplib2==0.11.1
idna==2.6
instamojo-wrapper==1.1
phonenumberslite==8.9.0
Pillow==5.0.0
PyPDF2==1.26.0
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
simplejson==3.13.2
six==1.11.0
swapper==1.1.0
urllib3==1.22
webencodings==0.5.1
xhtml2pdf==0.2.1
